+++ src/tbx/mod.rs
@@ -117,11 +117,11 @@ pub struct Reader {
     itr: Option<*mut htslib::hts_itr_t>,
 
     /// The currently fetch region's tid.
-    tid: i32,
+    tid: i64,
     /// The currently fetch region's 0-based begin pos.
-    start: i32,
+    start: i64,
     /// The currently fetch region's 0-based end pos.
-    end: i32,
+    end: i64,
 }
 
 unsafe impl Send for Reader {}
@@ -209,10 +209,10 @@ impl Reader {
     }
 
     /// Fetch region given by numeric sequence number and 0-based begin and end position.
-    pub fn fetch(&mut self, tid: u32, start: u32, end: u32) -> Result<()> {
-        self.tid = tid as i32;
-        self.start = start as i32;
-        self.end = end as i32;
+    pub fn fetch(&mut self, tid: u32, start: u64, end: u64) -> Result<()> {
+        self.tid = tid as i64;
+        self.start = start as i64;
+        self.end = end as i64;
 
         if let Some(itr) = self.itr {
             unsafe {
@@ -278,7 +278,7 @@ impl Reader {
 }
 
 /// Return whether the two given genomic intervals overlap.
-fn overlap(tid1: i32, begin1: i32, end1: i32, tid2: i32, begin2: i32, end2: i32) -> bool {
+fn overlap(tid1: i64, begin1: i64, end1: i64, tid2: i64, begin2: i64, end2: i64) -> bool {
     (tid1 == tid2) && (begin1 < end2) && (begin2 < end1)
 }
 
@@ -310,7 +310,8 @@ impl Read for Reader {
                     // returns `< 0`).
                     let (tid, start, end) =
                         unsafe { ((*itr).curr_tid, (*itr).curr_beg, (*itr).curr_end) };
-                    if overlap(self.tid, self.start, self.end, tid,start,end) {
+                    // XXX: Careful with this tid conversion!!!
+                    if overlap(self.tid, self.start, self.end, tid as i64,start,end) {
                         *record =
                             unsafe { Vec::from(ffi::CStr::from_ptr(self.buf.s).to_str().unwrap()) };
                         return Ok(true);
